src/requirements.txt
# Core Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Web App (Optional)
streamlit>=1.25.0

# Google Colab (Optional)
# google-colab

# Utilities
tqdm>=4.65.0